requirements.txt
toml
tqdm
pydub
jinja2
pillow
requests
boto3==1.35.99
arxiv
google-generativeai>=0.8.3
beautifulsoup4
pdf2image
paperswithcode-client
openreview-py
